Basic function set (read description)
client.html and client.js are example files for the respective implementation of 'Dozent' and 'Student'. The Files hold the html script and js code. Only functions specific to that client are here, for example handling input and keeping date up to date.\n\nfunctions.js is a general set of functions to be used to interact with the database. I will probably add a function as in the wizard to help keep data up to date, but there should not be much more to come.\n\ncouchdb.js handels the communication with the database. Your username and password is expected to admin and admin and the local port being default on 127.0.0.1:5984. For ease of use, these parameters are not to changed.